change the data type of bit rate from int to double for supporting decimal bit rates

+++ trunk/csharp/daisy/export/Daisy3_Export_ImageDescription_BodyContent.cs
@@ -648,7 +648,7 @@ namespace urakawa.daisy.export
             string imageDescriptionDirectoryPath,
             Dictionary<AlternateContentProperty, Description> map_AltProperty_TO_Description,
             bool encodeToMp3,
-            int bitRate_Mp3,
+            double bitRate_Mp3,
             AudioLibPCMFormat pcmFormat,
             Dictionary<AlternateContent, string> map_AltContentAudio_TO_RelativeExportedFilePath
             )
@@ -1252,7 +1252,7 @@ namespace urakawa.daisy.export
                     || name == DiagramContentModelHelper.D_Summary);
         }
 
-        private static string EncodeWavFileToMp3(bool skipACM, string sourceFilePath, bool encodeToMp3, AudioLibPCMFormat pcmFormat, int bitRate_Mp3)
+        private static string EncodeWavFileToMp3(bool skipACM, string sourceFilePath, bool encodeToMp3, AudioLibPCMFormat pcmFormat, double bitRate_Mp3)
         {
             AudioLib.WavFormatConverter formatConverter = new WavFormatConverter(true, skipACM);
